Move Table to util package
